feat(admin):删除评论

+++ claw-admin/claw-admin-admin/src/main/java/com/ruoyi/web/controller/content/ActivityController.java
@@ -6,6 +6,7 @@ import com.ruoyi.common.core.controller.BaseController;
 import com.ruoyi.common.core.domain.AjaxResult;
 import com.ruoyi.common.core.domain.R;
 import com.ruoyi.common.core.page.TableDataInfo;
+import com.ruoyi.common.exception.ServiceException;
 import com.ruoyi.common.utils.SecurityUtils;
 import com.ruoyi.system.service.ActivityService;
 import com.ruoyi.system.service.FileUploadService;
@@ -49,6 +50,9 @@ public class ActivityController extends BaseController {
      */
     @DeleteMapping("/{id}")
     public AjaxResult deleteActivity(@PathVariable("id") String id) {
+        if (id == null) {
+            throw new ServiceException("请选择要删除的评论");
+        }
         List<String> ids = Arrays.asList(id.split(","));
         activityService.deleteActivityByIds(ids);
         return success();
